Change the data frame slightly and added two files with constructed models
The change in the data frame only pertains to the names of columns such as spelling mistakes and large letters. The file data_prep is also changed per the above changes.

Two models files are added. The full_model file is for two cases of reducing a full model to reduced models. The first includes the variables Wealthy and Municipality and the second works with the variable Parish instead.

The other file model_restricted_municipality restricts the data frame to each of the four municipalities and then does the same analysis as full_model.